+++ internal/frontend/qml/tst_Gui.qml
@@ -122,7 +122,6 @@ Window {
         ListElement { title: "Minimize this"  }
         ListElement { title: "SendAlertPopup" }
         ListElement { title: "TLSCertError"   }
-        ListElement { title: "IMAPCertError"  }
     }
 
     ListView {
@@ -208,9 +207,6 @@ Window {
                     break;
                     case "TLSCertError" :
                     go.showCertIssue()
-                    break;
-                    case "IMAPCertError" :
-                    go.showIMAPCertTroubleshoot()
                     break;
                     default :
                     console.log("Not implemented " + data)
@@ -314,7 +310,6 @@ Window {
         signal failedAutostartCode(string code)
 
         signal showCertIssue()
-        signal showIMAPCertTroubleshoot()
 
         signal updateFinished(bool hasError)